% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA2.0.R
\name{chopString}
\alias{chopString}
\title{Chop string}
\usage{
chopString(string, nChar = 10, add.dots = TRUE)
}
\arguments{
\item{string}{character}

\item{nChar}{numeric}

\item{add.dots}{logical}
}
\value{
string
}
\description{
Given a string and a number of characters, the function chops the string to the
specified number of characters and adds '...' to the end.
}
\keyword{internal}
