% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA2-fx.R
\name{gsea_score}
\alias{gsea_score}
\title{Calculate GSEA enrichment score}
\usage{
gsea_score(
  ordered.gene.list,
  gene.set2,
  weight,
  n.rows,
  correl.type,
  gene.set.direction,
  data.expr,
  statistic,
  min.overlap
)
}
\arguments{
\item{ordered.gene.list}{list}

\item{gene.set2}{TODO}

\item{weight}{integer}

\item{n.rows}{integer}

\item{correl.type}{character, one of "rank", "symm.rank", "z.score"}

\item{gene.set.direction}{character vector, "d" for down and "u" for up}

\item{data.expr}{TODO}

\item{statistic}{TODO}

\item{min.overlap}{TODO}
}
\value{
named list
}
\description{
Apply correlation scheme and weighting; calculate ES
}
\keyword{internal}
