% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA2.0.R
\name{my.col2rgb}
\alias{my.col2rgb}
\title{Translate a color name into RGB space}
\usage{
my.col2rgb(color, alpha = 80, maxColorValue = 255)
}
\arguments{
\item{color}{character vector}

\item{alpha}{integer}

\item{maxColorValue}{integer}
}
\value{
vector of results (one per input color)
}
\description{
Translate a color name into RGB space
}
\keyword{internal}
