% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA2-fx.R
\name{project_geneset}
\alias{project_geneset}
\title{Project geneset}
\usage{
project_geneset(
  data.array,
  gene.names,
  n.cols,
  n.rows,
  weight = 0,
  statistic = "Kolmogorov-Smirnov",
  gene.set,
  nperm = 200,
  correl.type = "rank",
  gene.set.direction = NULL,
  min.overlap,
  size.G.current
)
}
\arguments{
\item{data.array}{TODO}

\item{gene.names}{TODO}

\item{n.cols}{TODO}

\item{n.rows}{TODO}

\item{weight}{TODO}

\item{statistic}{character, one of "Kolmogorov-Smirnov", "area.under.RES"}

\item{gene.set}{TODO}

\item{nperm}{integer, number of permutations}

\item{correl.type}{character, one of "rank", "z.score", "symm.rank"}

\item{gene.set.direction}{direction of regulation; has to be in same order as \code{gene.set}}

\item{min.overlap}{TODO}

\item{size.G.current}{TODO}
}
\value{
named list
}
\description{
Executed per gene set.
}
\keyword{internal}
