% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA2.0.R
\name{Read.GeneSets.db2}
\alias{Read.GeneSets.db2}
\title{Read gene sets}
\usage{
Read.GeneSets.db2(gs.db, thres.min = 2, thres.max = 2000)
}
\arguments{
\item{gs.db}{}

\item{thres.min}{}

\item{thres.max}{}
}
\value{
named list
}
\description{
Import gene sets into R workspace
}
\keyword{internal}
