% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA2-fx.R
\name{read_genesets_db2}
\alias{read_genesets_db2}
\title{Read gene set database}
\usage{
read_genesets_db2(gs.db, thres.min = 2, thres.max = 2000)
}
\arguments{
\item{gs.db}{path to gene database file in gmt/gct format}

\item{thres.min}{integer}

\item{thres.max}{integer}
}
\value{
named list
}
\description{
Read gene set database
}
\keyword{internal}
