% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA2-fx.R
\name{score}
\alias{score}
\title{Calculate ES score}
\usage{
score(max.ES, min.ES, RES, gaps, valleys, statistic)
}
\arguments{
\item{max.ES}{numeric}

\item{min.ES}{numeric}

\item{RES}{numeric vector}

\item{gaps}{numeric vector?}

\item{valleys}{numeric vector?}

\item{statistic}{character, "Kolmogorov-Smirnov" or "area.under.RES"}
}
\value{
named list:
\describe{
\item{RES}{TODO}
\item{ES}{TODO}
\item{arg.ES}{TODO}
}
}
\description{
Calculate ES score
}
\keyword{internal}
