% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA2-package.R
\docType{package}
\name{ssGSEA2-package}
\alias{ssGSEA2}
\alias{ssGSEA2-package}
\title{ssGSEA2: ssGSEA and PTM-SEA}
\description{
Single sample Gene Set Enrichment analysis (ssGSEA) and PTM Enrichment Analysis (PTM-SEA).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/broadinstitute/ssGSEA2.0}
  \item \doi{10.1074/mcp.TIR118.000943}
}

}
\author{
\strong{Maintainer}: Nicole Gay \email{nicole.r.gay@gmail.com}

Authors:
\itemize{
  \item Karsten Krug
  \item Pablo Tamayo
  \item The Broad Institute, Inc. [copyright holder]
}

Other contributors:
\itemize{
  \item National Cancer Institute Clinical Proteomic Tumor Analysis Consortium (CPTAC) [funder]
}

}
\keyword{internal}
