% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA2.0.R
\name{run_ssGSEA2}
\alias{run_ssGSEA2}
\title{Single sample GSEA}
\usage{
run_ssGSEA2(
  input.ds,
  output.prefix,
  gene.set.databases,
  sample.norm.type = c("rank", "log", "log.rank", "none"),
  weight = 0,
  statistic = c("area.under.RES", "Kolmogorov-Smirnov"),
  output.score.type = c("NES", "ES"),
  nperm = 1000,
  combine.mode = c("combine.off", "combine.replace", "combine.add"),
  min.overlap = 10,
  correl.type = c("rank", "z.score", "symm.rank"),
  global.fdr = FALSE,
  extended.output = TRUE,
  par = FALSE,
  spare.cores = 1,
  export.signat.gct = TRUE,
  param.file = TRUE,
  log.file = "run.log"
)
}
\arguments{
\item{input.ds}{Input data file in gct format, first column (Name) must contain gene symbols}

\item{output.prefix}{Prefix used for output tables}

\item{gene.set.databases}{List of genesets (in gmt format) to evaluate enrichment on}

\item{sample.norm.type}{Sample normalization, one of "rank", "log", "log.rank", "none"}

\item{weight}{When weight==0, all genes have the same weight; if weight>0,
actual values matter, and can change the resulting score. Default: 0}

\item{statistic}{Test statistic, one of "area.under.RES", "Kolmogorov-Smirnov"}

\item{output.score.type}{One of "NES", "ES"}

\item{nperm}{Number of random permutations for NES case. Default: 1000}

\item{combine.mode}{If "combine.off", do not combine \code{*_UP} and \code{*_DN} versions in a single score.
If "combine.replace", combine \code{*_UP} and \code{*_DN} versions in a single score.
If "combine.add" combine \code{*_UP} and \code{*_DN} versions in a single score and
add it but keeping the individual \code{*_UP} and \code{*_DN} versions.}

\item{min.overlap}{Default: 10}

\item{correl.type}{Correlation type: "rank", "z.score", "symm.rank"}

\item{global.fdr}{If TRUE calculate global FDR; else calculate FDR sample-by-sample. Default: FALSE}

\item{extended.output}{If TRUE the result GCT files will contain statistics about gene coverage etc. Default: TRUE}

\item{par}{Default: FALSE}

\item{spare.cores}{Default: spare.cores}

\item{export.signat.gct}{If TRUE gct files with expression values for each signature will be generated. Default: TRUE}

\item{param.file}{Default: TRUE}

\item{log.file}{Default: 'run.log'}
}
\value{
results
}
\description{
Single sample GSEA
}
\details{
For results similar to the Java version, use: weight=0;
when weight==0, sample.norm.type and correl.type do not matter;
when weight > 0, the combination of sample.norm.type and correl.type
dictate how the gene expression values in input.ds are transformed
to obtain the score -- use this setting with care (the transformations
can skew scores towards +ve or -ve values)
\itemize{
\item sample.norm.type=='none' uses actual expression values; combined
with correl.type=='rank', genes are weighted by actual values
\item sample.norm.type=='rank' weights genes proportional to rank
\item sample.norm.type=='log' can be used for log-transforming input data
\item correl.type=='z.score' standardizes the (normalized) input values
before using them to calculate scores.
}
}
